Moved stack widget package.
