+++ app/src/main/res/xml/stack_widget_provider.xml
@@ -9,4 +9,4 @@
     android:updatePeriodMillis="86400000"
     android:widgetCategory="keyguard|home_screen"
     android:previewImage="@drawable/widget_stack_preview"
-    android:configure="it.niedermann.nextcloud.deck.ui.appwidget.StackWidgetConfigurationActivity" />
+    android:configure="it.niedermann.nextcloud.deck.ui.widget.stack.StackWidgetConfigurationActivity" />
